requirements-dev.txt
python-dotenv==1.1
python-telegram-bot==22
google-adk==0.1.*
litellm==1.65.*